Break testusml.py up into multiple tests.
